chore(calendar): calendar 로직을 record 로직으로부터 분리 (#29)
